# schemas/api/openapi.yaml
openapi: 3.0.3
info:
  title: DIDHub API
  version: 0.1.0
  description: >-
    HTTP API surface for the DIDHub backend. This schema intentionally focuses on
    contracts needed by the generator stubs and TypeScript client.
servers:
  - url: https://api.didhub.local
    description: Placeholder server URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
paths:
...
  /files/content/{fileId}:
    get:
      tags: [Uploads]
      summary: Serve raw stored file content (bytes)
      operationId: serveStoredFileContent
      x-handler:
        delegate: crate::handlers::uploads::serve::serve_stored_file_content
        passHeaders: true
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: w
          in: query
          description: Desired width in pixels for an image thumbnail. If provided, server will return a resized thumbnail (maintains aspect ratio).
          required: false
          schema:
            type: integer
            minimum: 1
        - name: h
          in: query
          description: Desired height in pixels for an image thumbnail. If provided with w, server will fit image into w x h preserving aspect ratio.
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Raw file bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
security:
  - bearerAuth: []